LayoutTests/inspector/canvas/requestContent.html

<!DOCTYPE html>
<html>
<head>
<script src="../../http/tests/inspector/resources/inspector-test.js"></script>
<script>
function load() {
    window.contextOnscreen = document.getElementById("onscreen").getContext("2d");
    window.contextOnscreen.fillRect(25, 25, 250, 100);

    window.contextOffscreen = document.createElement("canvas").getContext("2d");
    window.contextOffscreen.fillRect(25, 25, 250, 100);

    window.contextCSS = document.getCSSCanvasContext("2d", "css", 10, 10);
    window.contextCSS.fillRect(2, 2, 6, 6);

    runTest();
}

function test() {
    let suite = InspectorTest.createAsyncSuite("Canvas.requestContent");

    suite.addTestCase({
        name: "Canvas.requestContent.validCanvasId",
        description: "Get the base64 encoded data for each canvas on the page.",
        test(resolve, reject) {
            let canvases = WebInspector.canvasManager.canvases;
            let expectedLength = canvases.length;
            InspectorTest.assert(expectedLength === 3, "The page has 3 canvases.");

            let contentsMap = new Map;

            function finish() {
                let results = {};
                for (let [canvas, content] of contentsMap)
                    results[canvas.displayName] = content;

                let keys = Object.keys(results);
                InspectorTest.assert(keys.length === contentsMap.size, "No display name collisions");

                // Ensure that the test runs properly even if the canvas
                // events are sent in a different order than the above.
                keys.sort();
                for (let displayName of keys) {
                    InspectorTest.log(`Canvas "${displayName}" has content:`);
                    InspectorTest.log(results[displayName]);
                    InspectorTest.log("");
                }

                resolve();
            }

            canvases.forEach((canvas) => {
                CanvasAgent.requestContent(canvas.identifier, (error, content) => {
                    if (error) {
                        reject(error);
                        return;
                    }

                    contentsMap.set(canvas, content);
                    if (contentsMap.size === expectedLength)
                        finish();
                });
            });
        }
    });

    // ------

    suite.addTestCase({
        name: "Canvas.requestContent.invalidCanvasId",
        description: "Invalid canvas identifiers should cause an error.",
        test(resolve, reject) {
            const canvasId = "DOES_NOT_EXIST";
            CanvasAgent.requestContent(canvasId, (error) => {
                InspectorTest.expectThat(error, "Should produce an error.");
                InspectorTest.log("Error: " + error);
                resolve();
            });
        }
    });

    suite.runTestCasesAndFinish();
}
</script>
<style>
    .css { background-image: -webkit-canvas(css); }
</style>
</head>
<body onload="load()">
<p>Test that CanvasAgent.requestContent can properly resolve the owner canvas node.</p>
<canvas id="onscreen"></canvas>
</body>
</html>
